requirements.txt
# Django core
Django==4.2.21
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2025.1

# Django extensions
django-allauth==65.2.0
django-axes==6.5.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-summernote==0.8.20.0
dj-database-url==2.3.0

# Deployment & static files
whitenoise==6.7.0
gunicorn==22.0.0

# Database driver
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.1
